apiFaceDog/requirements.txt
anyio==3.3.4
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
dnspython==2.1.0
email-validator==1.1.3
et-xmlfile==1.1.0
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
ldap3==2.9.1
numpy==1.22.2
openpyxl==3.0.9
pandas==1.4.0
passlib==1.7.4
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
pymongo==3.12.1
PyMySQL==1.0.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.1
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
requests==2.27.1
schedule==1.1.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.16.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.8
uvicorn==0.15.0
voluptuous==0.12.2
watchgod==0.7
websockets==10.0
